Replace mktemp() with mkdtemp() and check error.
The use of mktemp() was triggering warnings, and the man page is scary.
The mktemp() created a unique directory name, the directory itself was
created in ctdir(). Note, however, that ctdir() didn't check for mkdir()
failure. The revised code *always* creates a directory instead of doing
it "on request" from the test case. That may be slower, but at least now
we'll notice if the directory couldn't be created.